requirements.txt
boto==2.21.0
python-gflags==2.0
wsgiref==0.1.2
